# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Councel Connect",
        "version": "1.0.0"
    },
    "paths": {
        "/api/user/skill": {
            "post": {
                "tags": [
                    "skill"
                ],
                "summary": "User skill",
                "description": "User skill here",
                "operationId": "skill",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "mobile",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "example": "shiva"
                                    },
                                    "email": {
                                        "type": "text",
                                        "example": "shiva9291@gmail.com"
                                    },
                                    "mobile": {
                                        "type": "number",
                                        "example": "9876543210"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "09876543210"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": "09876543210"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "User Register",
                "description": "User Register here",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "mobile",
                                    "password",
                                    "password_confirmation",
                                    "_token"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "shiva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "shiva9291@gmail.com"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "09876543210"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "09876543210"
                                    },
                                    "_token": {
                                        "description": "CSRF token. Include this as a hidden field in the form or as a header.",
                                        "type": "string",
                                        "example": "csrf_token_here"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    }
}